支持K系列的OTA功能，2.2.2.15

+++ iOS_sdk_source/Demo/VeepooBleSDKDemo/VeepooBleSDK.framework/Headers/VPDFUOperation.h
@@ -16,7 +16,7 @@
  @param progress 升级进度
  @param deviceDFUState 升级的状态变化
  */
-typedef void(^DeviceDFUBlock)(NSInteger progress, DeviceDFUState deviceDFUState);
+typedef void(^DeviceDFUBlock)(double progress, DeviceDFUState deviceDFUState);
 
 @interface VPDFUOperation : NSObject
